style: 컬럼 확장을 위한 모듈화
