+++ .gitignore
@@ -1,6 +1,8 @@
 .DS_Store
 
 node_modules
+/node_modules
+node_modules/*
 dist
 dist-ssr
 *.local
