+++ src/constant/index.ts
@@ -1,0 +1,1 @@
+export const GRID = 8;
